RELATED: BB-1612 Refactor PivotTable sticky row
