+++ src/components/core/pivotTable/stickyRowHandler.ts
@@ -5,11 +5,16 @@ import { colIdIsSimpleAttribute, getGridIndex } from "./agGridUtils";
 import ApiWrapper from "./agGridApiWrapper";
 import { getScrollbarWidth } from "../../../helpers/domUtils";
 
-export const initStickyHeaders = (gridApi: GridApi) => {
+export interface IScrollPosition {
+    readonly top: number;
+    readonly left: number;
+}
+
+export const initializeStickyRow = (gridApi: GridApi) => {
     gridApi.setPinnedTopRowData([{}]);
 };
 
-export const updateStickyHeadersPosition = (gridApi: GridApi, apiWrapper: any = ApiWrapper) => {
+export const updateStickyRowPosition = (gridApi: GridApi, apiWrapper: any = ApiWrapper) => {
     const headerHeight = apiWrapper.getHeaderHeight(gridApi);
     apiWrapper.setPinnedTopRowStyle(gridApi, "top", `${headerHeight}px`);
     apiWrapper.setPinnedTopRowStyle(gridApi, "padding-right", `${getScrollbarWidth()}px`);
@@ -20,40 +25,39 @@ export function stickyRowExists(gridApi: GridApi, apiWrapper: any = ApiWrapper):
 }
 
 function shouldUpdate(
-    currentScrollTop: number,
-    currentScrollLeft: number,
-    lastScrollTop: number,
-    lastScrollLeft: number,
+    currentScrollPosition: IScrollPosition,
+    lastScrollPosition: IScrollPosition,
     rowHeight: number,
 ) {
-    const initialUpdate = currentScrollTop === 0;
-    const currentRowIndex = getGridIndex(currentScrollTop, rowHeight);
-    const lastRowIndex = getGridIndex(lastScrollTop, rowHeight);
+    const initialUpdate = currentScrollPosition.top === 0;
+    const currentRowIndex = getGridIndex(currentScrollPosition.top, rowHeight);
+    const lastRowIndex = getGridIndex(lastScrollPosition.top, rowHeight);
     const differentRow = currentRowIndex !== lastRowIndex;
     // when scrolling horizontally update with the same cadence as rows as we don't know where the column borders are
     const horizontalBreakpointDistance = rowHeight;
-    const currentHorizontalBreakpoint = getGridIndex(currentScrollLeft, horizontalBreakpointDistance);
-    const lastHorizontalBreakpoint = getGridIndex(lastScrollLeft, horizontalBreakpointDistance);
+    const currentHorizontalBreakpoint = getGridIndex(
+        currentScrollPosition.left,
+        horizontalBreakpointDistance,
+    );
+    const lastHorizontalBreakpoint = getGridIndex(lastScrollPosition.left, horizontalBreakpointDistance);
     const differentHorizontalBreakpoint = currentHorizontalBreakpoint !== lastHorizontalBreakpoint;
 
     return initialUpdate || differentRow || differentHorizontalBreakpoint;
 }
 
-export const updateStickyHeaders = (
-    currentScrollTop: number,
-    currentScrollLeft: number,
-    lastScrollTop: number,
-    lastScrollLeft: number,
+export const updateStickyRowContentClasses = (
+    currentScrollPosition: IScrollPosition,
+    lastScrollPosition: IScrollPosition,
     rowHeight: number,
     gridApi: GridApi,
     groupingProvider: IGroupingProvider,
     apiWrapper: any,
 ) => {
-    if (!shouldUpdate(currentScrollTop, currentScrollLeft, lastScrollTop, lastScrollLeft, rowHeight)) {
+    if (!shouldUpdate(currentScrollPosition, lastScrollPosition, rowHeight)) {
         return;
     }
 
-    const firstVisibleRowIndex = getGridIndex(currentScrollTop, rowHeight);
+    const firstVisibleRowIndex = getGridIndex(currentScrollPosition.top, rowHeight);
     const firstVisibleRow: RowNode = gridApi.getDisplayedRowAtIndex(firstVisibleRowIndex);
     const firstVisibleNodeData = firstVisibleRow && firstVisibleRow.data ? firstVisibleRow.data : null;
 
@@ -63,7 +67,7 @@ export const updateStickyHeaders = (
     }
     apiWrapper.addPinnedTopRowClass(gridApi, "gd-visible-sticky-row");
 
-    const lastRowIndex = getGridIndex(lastScrollTop, rowHeight);
+    const lastRowIndex = getGridIndex(lastScrollPosition.top, rowHeight);
     const attributeKeys = Object.keys(firstVisibleNodeData).filter(colIdIsSimpleAttribute);
 
     attributeKeys.forEach((columnId: string) => {
@@ -81,7 +85,7 @@ export const updateStickyHeaders = (
             // if the column has some groups
             if (groupingProvider.isColumnWithGrouping(columnId)) {
                 // show the last cell of the group temporarily so it scrolls out of the viewport nicely
-                const currentRowIndex = getGridIndex(currentScrollTop, rowHeight);
+                const currentRowIndex = getGridIndex(currentScrollPosition.top, rowHeight);
                 apiWrapper.addCellClass(gridApi, columnId, currentRowIndex, "gd-cell-show-hidden");
             }
         }
